VC5 compat fixes + relocation-aware matching: 74 -> 75 matches
VC5 compile fixes (0 failures now, was 21):
- br_ui.h: disable negative-subscript static asserts under VC5
- slice2_16.c: long long -> int64_t (via stdint.h shim)
- slice2_24.c: C99 designated initializers -> positional
- slice7_82.c: guard POSIX headers/functions with #ifndef _MSC_VER
- br_drawenv.c: move mid-block declarations to block start
- br_sfx.c: avoid unsigned __int64 -> double cast

match_diff.py: read COFF relocation tables and mask relocated bytes
(call/jmp targets, global addresses) during comparison. This eliminated
false diffs from linker-resolved addresses and jumped the match count
from 29 to 74.

slice1_02.c: BrFixUnpackU32Q13 matched by using float constant (d8
fmul) instead of double (dc fmul). Same pattern identified in the
other BrFixUnpack functions but those need deeper structural changes.

102/102 source files compile under VC5. 136/136 port tests pass.
75 functions match byte-identical (deduplicated across /O2 and /Od).

+++ src/core/audio/br_sfx.c
@@ -314,7 +314,7 @@ int64_t BrSfxRatioFromHz(uint32_t hz, double baseRate)
      * high dword zeroed and loaded with `fild qword`, so it is
      * zero-extended, then multiplied by 0x10077C08 (2^32) and divided by the
      * channel's base rate.  No guard on baseRate. */
-    return br_ftol64((double)(uint64_t)hz * 4294967296.0 / baseRate);
+    return br_ftol64((double)(int64_t)(uint32_t)hz * 4294967296.0 / baseRate);
 }
 
 uint32_t BrSfxHzFromRatio(int64_t ratio, double baseRate)
